requirements.txt
huggingface_hub>=0.20.0,<1.0.0
hf_transfer
hf_xet
pillow>=9.0.0
numpy>=1.21.0
trimesh>=3.0.0
imageio>=2.0.0
imageio-ffmpeg>=0.4.0
easydict
plyfile
zstandard
opencv-python-headless
tqdm
transformers>=4.56.0
kornia
timm
lpips
safetensors
einops
gguf
psutil
